Organize crate::tracing better
